rag-langchain-milvus/pyproject.toml
[project]
name = "rag-langchain-milvus"
version = "0.1.0"
description = ""
authors = [
    {name = "balajihambeere",email = "balaji.hambere@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "langchain (>=0.3.19,<0.4.0)",
    "langchain-openai (>=0.3.7,<0.4.0)",
    "langchain-community (>=0.3.18,<0.4.0)",
    "langchain-text-splitters (>=0.3.6,<0.4.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "langchain-milvus (>=0.1.8,<0.2.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
